backend/requirements.txt
# CyberVision Backend - Python Dependencies
# FastAPI Framework

fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Image Processing
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==1.26.4

# EXIF Metadata
piexif==1.1.3

# File Type Detection
python-magic==0.4.27

# Statistical Analysis
scipy==1.14.1

# Environment Variables
python-dotenv==1.0.1

# HTTP Client
requests==2.32.3

# CORS
# (inclus avec FastAPI mais explicit ici)

# Optional: Development Tools
# pytest==8.3.4
# black==24.10.0
# flake8==7.1.1